chore(): export additional components & types
