+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@awell_health/ui-library",
-  "version": "0.0.50",
+  "version": "0.0.51",
   "private": false,
   "description": "UI components to integrate with Awell Health",
   "repository": {
