+++ src/hostedPages/activities/wizardForm/index.ts
@@ -1,2 +1,1 @@
 export { WizardForm } from './WizardForm'
-export type { WizardFormProps } from './types'
