+++ src/hostedPages/index.tsx
@@ -2,3 +2,4 @@ export { Checklist } from './activities/checklist'
 export { Message } from './activities/message'
 export { WizardForm } from './activities/wizardForm'
 export { HostedPageLayout } from './layouts/HostedPageLayout'
+export { CloseButton } from './layouts/HostedPageLayout'
