+++ src/hostedPages/layouts/HostedPageLayout/index.tsx
@@ -1,1 +1,2 @@
 export { HostedPageLayout } from './HostedPageLayout'
+export { CloseButton } from './CloseButton'
